qa: fix P2-5 price alert review - NUMERIC precision, missing migration, Ice comment
- Fix price_level column from Float to NUMERIC(18,8) in PriceAlertORM for
  financial precision consistency with all other price columns in the system
- Add missing Alembic migration 20260220_create_price_alerts to create the
  price_alerts table (was absent - table would never be created in production)
- Fix WyckoffLevelType.ICE inline comment: was "support line after distribution"
  (incorrect context) → "Resistance/support line within accumulation range
  (becomes support after SOS breakout)" per lps_detector.py domain definition

+++ backend/src/orm/models.py
@@ -1351,8 +1351,9 @@ class PriceAlertORM(Base):
     )
 
     # Price level (NULL for phase_change alerts)
-    price_level: Mapped[float | None] = mapped_column(
-        Float,
+    # Use NUMERIC for financial precision (avoids floating-point rounding)
+    price_level: Mapped[Decimal | None] = mapped_column(
+        NUMERIC(18, 8),
         nullable=True,
     )
